watchdog: imxulp: Fix prescaler setting on i.MX7ulp
On i.MX7ulp the driver assumes a clock tick rate of 1kHz. This is only
true though when the /256 prescaler is disabled. To fix this we disable
the prescaler on i.MX7ulp while keeping it enabled on i.MX93. We derived
this bug from U-Boot, the Kernel does this correctly and has the same
code as we introduce here now.

Link: https://lore.barebox.org/20240115155615.953954-2-s.hauer@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>