ARM: i.MX8M: remove unutilized start parameter to i.MX8M load functions
We have two kinds of xload functions for i.MX: _start_image loads and
starts the image and _load_image takes a parameter on whether to start
the image after loading. All users of _load_image pass false as argument
for the start parameter and this is unlikely to change, because we need
to do other things before starting the image:

  - BL31: Install ARM Trusted Firmware
  - BL32: Optionally, extract OP-TEE from external firmware
  - BL33: Optionally, verify barebox proper when secure booting

_start_image doesn't concern itself with this:

  - BL31: We don't use TF-A on 32-bit i.MX systems and if barebox is secure
    monitor, it's installed in barebox proper
  - BL32: OP-TEE is shipped inside PBL or loaded by bootm on 32-bit i.MX
  - BL33: barebox is loaded directly to SDRAM after it was setup by DCD
          and is thus verified by bootrom as a whole. Chainloading
	  manually is not mixed with secure boot

For that reason, let's drop the start parameter to try to reduce
the complexity of the chainloading code a bit.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240115170110.321676-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>