ARM: i.MX8M: return error if imx_load_image can't honour entry address
On i.MX6 and 7, we call imx_load_image with start == true and address
equal to entry.

On i.MX8M, we call imx_load_image with start == false and address
unequal to entry.

imx_load_image interprets the address being unequal to entry as a
directive to move the image, so that the barebox entry point without
the i.MX header starts at exactly the entry address.

If we were to change this, say on an i.MX8MN, so address is equal
to entry, the system will no longer boot:

  - i.MX header is loaded to offset 0 from start of SDRAM (like on i.MX6/7)
  - barebox PBL entry point is loaded to offset 32K
  - imx8mX_load_bl33 will copy the running barebox PBL to offset 0

The result of this is that the TF-A will be able to return to barebox
(non-executable i.MX header overwritten with executable PBL), but
uncompressing barebox will fail (remainder of partially overwritten
PBL is interpreted as start of piggydata).

Add some documentation explaining this complexity, a hint on how we
could improve it and change the condition, so entry == address results
in an explicit warning message immediately.

Reported-by: Marco Felsch <m.felsch@pengutronix.de>
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240115170110.321676-2-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>