firmware: fix external firmware dependencies
A pbl is not rebuilt when the firmware file specified in fw-external-y
changes.

$(patsubst %.gen.o,$(obj)/%.extgen.pbl.o, $(pbl-y)) is executed before
pbl-y actually contains the contents of fw-external-y. Also we want to
do the subsitution only for the external firmware files, not for pbl-y.

To fix this use a intermediate variable pbl-fwext-y instead of pbl-y and
initialize it before usage.

Also, use $$(wildcard $(fwdir)/%) instead of $(fwdir)/% to successfully
compile when the external firmware file is missing.

Reviewed-by: Marco Felsch <m.felsch@pengutronix.de>
Reviewed-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240116091020.2971875-1-s.hauer@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>