ARM: i.MX8M: fix optee of-fixup logic
The current code checks only if "/firmware/optee" exist on the builtin
dtb and applys the fixup if not found and if found nothing is done. If a
builtin dts contains the node but an external don't the fixup won't be
applied. Also if the external dts does have a node + the reserved memory
region nodes but the barebox builtin dts don't we do add additional
reserved memory nodes which may conflict due to different name scheme:
<name> vs. <name>@<addr>.

Move the "/firmware/optee" check into the of_optee_fixup() so the check
is done on the correct dtb root nodes.

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>
Link: https://lore.barebox.org/20240116170738.209954-19-m.felsch@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>