pmdomain: power: drop unused parameters for internal functions
We don't do latency measurements for power domain operations in barebox,
so the timed parameter to indicate whether time measurements should be
done or not is unused. We also don't need locking for the powerdomains,
so maintaining the depth for lockdep is also unnecessary.

Drop both unused function parameters as we are not likely to ever use
them.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240119162610.1014870-4-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>