pmdomain: drop superfluous parameter to __genpd_dev_pm_attach
The Linux prototype of __genpd_dev_pm_attach has the base device as
second argument and retrieves the of_node pointer via the device in the
first argument. As the function isn't exported, we don't have to follow
the Linux API, but we also don't need to keep arguments we don't use,
thus drop the device pointer.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240119162610.1014870-7-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>