driver: have dev_request_mem_region_err_null warn if resource starts at 0
dev_request_mem_region_err_null() is used very scarcely, because it
doesn't allow differentiating between NULL and a MMIO region at address 0.

Using it in the latter case is always a bug, so add a WARN_ON that warns
about this. This should make it safe to use at more places, which will
come in handy to implement Linux API that uses NULL as error value.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240119162610.1014870-14-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>