ARM: dts: i.MX8MP: drop barebox,allow-dummy for HSIO blk-ctrl
Now that we have a HSIO blk ctrl driver, there is no need to generally
assume that TF-A or BootROM have left them enabled for us.

Boards that want to continue making use of this assumption can reinstate
the barebox,allow-dummy property, but everyone else should just enable
CONFIG_IMX8M_BLK_CTRL.

While barebox,allow-dummy only has an effect if the driver doesn't
exist (as determined by deep probe), we remove the property for improved
user experience when the TF-A doesn't enable the HSIO power domain.

In that case, barebox,allow-dummy would cause barebox to continue
probe of USB controller and PHY resulting in a bus hang instead of
an error message that tells what's up:

  imx8mq-usb-phy 381f0040.usb-phy@381f0040.of: error No such device: power domain attach failed
  dwc3-of-simple 32f10100.usb@32f10100.of: error No such device: power domain attach failed
  imx8mq-usb-phy 382f0040.usb-phy@382f0040.of: error No such device: power domain attach failed
  dwc3-of-simple 32f10108.usb@32f10108.of: error No such device: power domain attach failed

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240119162610.1014870-20-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>