ARM: i.MX93: tqm93xx: add PMIC node
The PMIC is needs to be configured on the tqm93xx to properly reset the
board, see the board specific initialisation code. We had a PMIC node
in our downstream dts file which we dropped in favour for using the
upstream dts. The upstream dts doesn't have a PMIC node yet, so add it
again. We do not need the regulator subnodes though, as barebox doesn't
support them anyway.

Link: https://lore.barebox.org/20240205125301.2465254-1-s.hauer@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>