serial: ns16550: add support for half duplex rs485
Some RS485 serial interfaces does not use correctly RTS as Receiver
Enable and Data Enable, the internal receiver of the ns16550 must be
disabled during transmit to avoid loopback.

By default the option is enabled (when the property
'linux,rs485-enabled-at-boot-time' is present), and can be disabled with
the property 'rs485-rx-during-tx'.

Signed-off-by: Pierre-Olivier Huard <pierre-olivier.huard@rtone.fr>
Link: https://lore.barebox.org/20240206152239.341460-3-pierre-olivier.huard@rtone.fr
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>