ARM: at91: sama5d27-som1: autoping watchdog during barebox runtime
The BootROM arms the watchdog with the maximum timeout of 16 seconds on
boot. This is ample time to boot into Linux userspace, so so far we
didn't touch the watchdog for the SAMA5D2 in barebox and expected the
user to disable the watchdog manually, e.g. via the `wd -x` command or
by storing nv.autoboot=abort into the environment.

While this works, this is admittedly bad user experience, so let's just
enable watchdog polling in barebox instead. That way, we still have
watchdog supervision, but don't reset when dropping into the barebox
shell.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240212154617.2260448-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>