driver: make CONFIG_DEBUG_PROBES more useful for removal
When CONFIG_DEBUG_PROBES is enabled, barebox will print a message on
every device probe and removal. Unfortunately, the removal prints are not
very useful, because the removal happens in the bus remove function,
which is often a no-op, but that's not known to driver core.

To make this a bit more useful, let's allow skipping bus remove
functions like Linux does and only print that a remove is happening if
either a bus or driver remove function is available.

At present, this doesn't change much, but the follow-up commit will drop
bus remove functions that only call the driver remove function, which
will shorten the CONFIG_DEBUG_PROBES output on shutdown a fair bit.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240215103353.2799723-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>