ARM: Layerscape: TQMLS1046a: switch to upstream DT
The existing DT was added alongside board support in v2019.05.0.
The board gained an upstream DT in Linux v6.7-rc1, which was pulled
as part of v2023.12.0. So let's drop our copy to avoid duplication.

This change will introduce user noticeable differences:

  - barebox compatible, which is also used to match kernel DT, changes
  - hostname changes as it's generated from DT compatible

This is deemed acceptable as it aligns us with the upstream DT.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240219133340.3873032-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>