treewide: use dev_err_probe instead of comparisons against EPROBE_DEFER
Drivers should not need to compare an error value against EPROBE_DEFER.
We have a number of drivers doing that though to decide whether to print
an error or not. This error message will be lost if the probe is
deferred, so use dev_err_probe to store the error in that case.

While at it, we shorten the error messages a bit. dev_err_probe will
already print the string 'error' before the error code string.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240219172659.3796647-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>